backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
python-dotenv==1.0.0
boto3==1.34.27
minio==7.2.3
spacy==3.7.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
